Fixed type of nbytes variable passed to ibrd(), which is probably causing problems on 64 bit systems.
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@1401 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/driver/sys/osfuncs.c
@@ -423,7 +423,7 @@ static int read_ioctl( gpib_file_private_t *file_priv, gpib_board_t *board,
 	int retval;
 	ssize_t read_ret = 0;
 	gpib_descriptor_t *desc;
-	int nbytes;
+	size_t nbytes;
 
 	retval = copy_from_user(&read_cmd, (void*) arg, sizeof(read_cmd));
 	if (retval)
